Initial role/playbook structure
This initial commit adds a basic playbook directory structure
for creating the role for bokeh.

For now, it just makes sure the needed dependencies are installed
and/or configured (e.g. apache modules, php5 modules, etc).

It also fetches the source code from the official Git repository and
checks the configured version. It defaults to the current 7.1 stable
release (7.1.16).

It is written so the user can parametrize its Bokeh deployment by setting
custom values in the vars/user.yml file. The default values are writen in
vars/defaults.yml and shouldn't be changed.

A small README.md is included.

It also includes a sample playbook (site.yml) so people can have an idea
of how to use it.

Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>